ci(bench): isolate files and use median samples
