+++ scripts/lib/median.ts
@@ -1,0 +1,8 @@
+export function median(values: readonly number[]): number {
+    if (values.length === 0) throw new Error("Cannot take median of no values")
+    const sorted = [...values].sort((a, b) => a - b)
+    const middle = Math.floor(sorted.length / 2)
+    return sorted.length % 2 === 0
+        ? (sorted[middle - 1] + sorted[middle]) / 2
+        : sorted[middle]
+}
